Fix: Output file not properly handled
The issue of output directory facing error occurred due to the assumption that the "OUTPUT_DIR" variable always contains the full path to the output directory. So, when specifying a subdirectory as part of the command-line argument for the output file (-o option), this assumption breaks.

With this modification, the code will properly handle the output directory structure, whether it's a full path or a subdirectory provided by the user as a command-line argument. This should resolve the "Directory not found" error.