Update app.py
In this version, I've added a number input widget to specify the number of exam centers to calculate. I've also modified the run_center_randomizer function to accept the number of centers as an argument and updated the subprocess command accordingly. This version provides more flexibility and control to the user.